+++ tests/core/timeseries/test_request.py
@@ -23,9 +23,9 @@ def expected_stations_df():
             (
                 "02480",
                 dt.datetime(2004, 9, 1, tzinfo=ZoneInfo("UTC")),
-                108.0,
                 50.0643,
                 8.993,
+                108.0,
                 "Kahl/Main",
                 "Bayern",
                 9.759384982994229,
@@ -33,9 +33,9 @@ def expected_stations_df():
             (
                 "04411",
                 dt.datetime(2002, 1, 24, tzinfo=ZoneInfo("UTC")),
-                155.0,
                 49.9195,
                 8.9672,
+                155.0,
                 "Schaafheim-Schlierbach",
                 "Hessen",
                 10.160326,
@@ -43,9 +43,9 @@ def expected_stations_df():
             (
                 "07341",
                 dt.datetime(2005, 7, 16, tzinfo=ZoneInfo("UTC")),
-                119.0,
                 50.0900,
                 8.7862,
+                119.0,
                 "Offenbach-Wetterpark",
                 "Hessen",
                 12.891318342515483,
@@ -53,10 +53,10 @@ def expected_stations_df():
         ],
         schema={
             "station_id": pl.Utf8,
-            "from_date": pl.Datetime(time_zone="UTC"),
-            "height": pl.Float64,
+            "start_date": pl.Datetime(time_zone="UTC"),
             "latitude": pl.Float64,
             "longitude": pl.Float64,
+            "height": pl.Float64,
             "name": pl.Utf8,
             "state": pl.Utf8,
             "distance": pl.Float64,
@@ -293,10 +293,10 @@ def test_dwd_observation_stations_filter_by_rank_single(default_request, expecte
         latlon=(50.0, 8.9),
         rank=1,
     )
-    given_df = request.df.drop(columns="to_date")
+    given_df = request.df.drop(columns="end_date")
     assert_frame_equal(given_df[0, :], expected_stations_df[0, :])
     values = request.values.all()
-    assert_frame_equal(values.df_stations[0, :].drop(columns="to_date"), expected_stations_df[0, :])
+    assert_frame_equal(values.df_stations.head(1).drop(columns="end_date"), expected_stations_df.head(1))
 
 
 @pytest.mark.remote
@@ -305,31 +305,31 @@ def test_dwd_observation_stations_filter_by_rank_multiple(default_request, expec
         latlon=(50.0, 8.9),
         rank=3,
     )
-    given_df = request.df.drop(columns="to_date")
+    given_df = request.df.drop(columns="end_date")
     assert_frame_equal(
         given_df.head(3),
         expected_stations_df,
     )
     values = request.values.all()
-    assert_frame_equal(values.df_stations.drop(columns="to_date"), expected_stations_df)
+    assert_frame_equal(values.df_stations.drop(columns="end_date"), expected_stations_df)
 
 
 @pytest.mark.remote
 def test_dwd_observation_stations_nearby_distance(default_request, expected_stations_df):
     # Kilometers
     nearby_station = default_request.filter_by_distance(latlon=(50.0, 8.9), distance=16.13, unit="km")
-    nearby_station = nearby_station.df.drop(columns="to_date")
+    nearby_station = nearby_station.df.drop(columns="end_date")
     assert_frame_equal(nearby_station, expected_stations_df)
     # Miles
     nearby_station = default_request.filter_by_distance(latlon=(50.0, 8.9), distance=10.03, unit="mi")
-    nearby_station = nearby_station.df.drop(columns="to_date")
+    nearby_station = nearby_station.df.drop(columns="end_date")
     assert_frame_equal(nearby_station, expected_stations_df)
 
 
 @pytest.mark.remote
 def test_dwd_observation_stations_bbox(default_request, expected_stations_df):
     nearby_station = default_request.filter_by_bbox(left=8.7862, bottom=49.9195, right=8.993, top=50.0900)
-    nearby_station = nearby_station.df.drop(columns="to_date")
+    nearby_station = nearby_station.df.drop(columns="end_date")
     assert_frame_equal(nearby_station, expected_stations_df.drop(columns=["distance"]))