+++ tests/provider/dwd/dmo/test_api.py
@@ -55,26 +55,26 @@ def test_dwd_dmo_stations(default_settings):
     given_df = stations.all().df
     assert not given_df.is_empty()
     assert given_df.select(pl.all().max()).to_dicts()[0] == {
-        "from_date": None,
-        "height": 4670.0,
+        "station_id": "Z949",
         "icao_id": "ZYTX",
+        "start_date": None,
+        "end_date": None,
         "latitude": 79.59,
         "longitude": 179.2,
+        "height": 4670.0,
         "name": "ZWOENITZ",
         "state": None,
-        "station_id": "Z949",
-        "to_date": None,
     }
     assert given_df.select(pl.all().min()).to_dicts()[0] == {
-        "from_date": None,
-        "height": -350.0,
+        "station_id": "01001",
         "icao_id": "AFDU",
+        "start_date": None,
+        "end_date": None,
         "latitude": -78.27,
         "longitude": -176.1,
+        "height": -350.0,
         "name": "16N55W",
         "state": None,
-        "station_id": "01001",
-        "to_date": None,
     }
     station_names_sorted = given_df.sort(pl.col("name").str.n_chars()).get_column("name").to_list()
     assert station_names_sorted[:5] == ["ELM", "PAU", "SAL", "AUE", "HOF"]