+++ tests/provider/dwd/road/test_data.py
@@ -12,20 +12,20 @@ def test_dwd_road_weather(parameter):
     request = DwdRoadRequest(parameter).filter_by_station_id("A006")
     item = request.to_dict()["stations"][0]
     assert item == {
-        "height": 2.0,
+        "station_id": "A006",
+        "start_date": None,
+        "end_date": None,
         "latitude": 54.8892,
         "longitude": 8.9087,
+        "height": 2.0,
         "name": "Boeglum",
+        "state": "SH",
         "road_name": "L5S",
         "road_sector": "2",
         "road_surface_type": 1,
         "road_surroundings_type": 2,
         "road_type": 1,
-        "state": "SH",
         "station_group": "KK",
-        "station_id": "A006",
-        "from_date": None,
-        "to_date": None,
     }
     values = (
         request.values.all().df.drop_nulls(subset="value").filter(pl.col("parameter").eq("temperature_air_mean_200"))