+++ wetterdienst/provider/dwd/dmo/api.py
@@ -1337,11 +1337,11 @@ class DwdDmoRequest(TimeseriesRequest):
     _base_columns = [
         Columns.STATION_ID.value,
         Columns.ICAO_ID.value,
-        Columns.FROM_DATE.value,
-        Columns.TO_DATE.value,
-        Columns.HEIGHT.value,
+        Columns.START_DATE.value,
+        Columns.END_DATE.value,
         Columns.LATITUDE.value,
         Columns.LONGITUDE.value,
+        Columns.HEIGHT.value,
         Columns.NAME.value,
         Columns.STATE.value,
     ]
@@ -1545,7 +1545,7 @@ class DwdDmoRequest(TimeseriesRequest):
             pl.col(Columns.ICAO_ID.value).map_dict({"----": None}, default=pl.col(Columns.ICAO_ID.value)),
             pl.col(Columns.LATITUDE.value).str.replace(" ", "").cast(float),
             pl.col(Columns.LONGITUDE.value).str.replace(" ", "").cast(float),
-            pl.lit(None, pl.Datetime(time_zone="UTC")).alias(Columns.FROM_DATE.value),
-            pl.lit(None, pl.Datetime(time_zone="UTC")).alias(Columns.TO_DATE.value),
+            pl.lit(None, pl.Datetime(time_zone="UTC")).alias(Columns.START_DATE.value),
+            pl.lit(None, pl.Datetime(time_zone="UTC")).alias(Columns.END_DATE.value),
             pl.lit(None, pl.Utf8).alias(Columns.STATE.value),
         )