+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Development
 ***********
 
+- [Breaking] Use start_date and end_date instead of from_date and to_date
+
 0.66.1 (08.11.2023)
 *******************